requirements.txt
Django==5.2
djangorestframework==3.16.0
mysqlclient==2.2.7
drf-spectacular==0.28.0
uvicorn==0.34.2
redis==6.0.0
celery==5.5.2
gunicorn==23.0.0